Created game display function, css grid display
Created a function to create buttons for every item
in the gameBoardArray. When button is clicked, turn result
determines whether a "X" or "O" is displayed on button which
is now disabled. Player's turn is also displayed and appended below
the title of the game.
Created CSS grid layout to replicate tic tac toe board.