Added if statement to fix turn display problem
The turn that was displayed and the X or O mark
didn't line up with the player array where the game
board number position was pushed into. I put an if
statement so the inital turn() call in the display
function returns the opposite result, so it will line
up with the turn that is called when a button is pressed.
Then added another if statement after button is pressed to
change turn display to the upcoming turn again.